pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "myfitnessbot"
version = "0.1.0"
description = "Telegram fitness bot with workouts, nutrition and payments"
authors = [{ name = "MyFitnessBot", email = "dev@myfitnessbot.local" }]
requires-python = ">=3.11"
dependencies = [
  "aiogram>=3.5.0",
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "httpx>=0.27.0",
  "apscheduler>=3.10.4",
  "Pillow>=10.4.0"
]

[project.optional-dependencies]
db = [
  "sqlalchemy>=2.0.30",
  "alembic>=1.13.2",
  "psycopg[binary,pool]>=3.1.19"
]
cache = [
  "redis>=5.0.4"
]
pdf = [
  "reportlab>=4.2.0",
  "weasyprint>=62.3",
  "jinja2>=3.1.4"
]
monitoring = [
  "sentry-sdk[fastapi]>=2.8.0",
  "prometheus-client>=0.20.0"
]
i18n = [
  "babel>=2.15.0"
]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.7",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
  "types-redis",
  "types-requests"
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["tests*", "data*", "tmp*", "scripts*"]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["app"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP", "B", "Q", "C4"]
ignore = ["E203", "E266", "E501", "W503"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false
warn_unused_ignores = true